Move server setup code to correct places
